--- src/components/Editor/plugins/NotesPlugin/index.ts
@@ -1,3 +1,0 @@
-import { NotesPlugin } from "./NotesPlugin";
-
-export default NotesPlugin;
