+++ src/components/Editor/plugins/NotesPlugin/utils/patches.ts
